Fixes regression relating to #4
For a has-many through relationship, A->B->C for which there is no
relationship C->B would, loading the relationship would omit records.
More specifically, the only records that would be loaded from table C
were those whose ID was equal to the ID of the A record.

The fix was to alter the eager loading logic in Relationship such that
records are not filtered by ID if the relationship is a has-many
through. I'm not 100% confident that this is the correct theoretical
solution, but it does pass all PHP AR and Zamzar API tests.